workmesh/pyproject.toml
# Copyright 2025 NVIDIA Corporation
# SPDX-License-Identifier: Apache-2.0

[project]
name = "workmesh"
version = "0.1.0"
description = "A python service library for Spark & Reachy Photo Booth"
readme = "README.md"
license = "Apache-2.0"
authors = [{ name = "NVIDIA Corporation" }]
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "confluent-kafka>=2.11.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "deepmerge>=2.0",
    "opentelemetry-distro>=0.58b0",
    "opentelemetry-instrumentation-aiokafka>=0.58b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.15.0",
    "grpcio>=1.67.1",
    "protobuf>=5.29.5",
]

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "grpcio-tools>=1.62.2",
]
